backend/requirements.txt
fastapi
uvicorn[standard]
httpx
vaderSentiment
motor
python-dotenv
pydantic
